requirements.txt
anyio==4.1.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
ephem==4.1.5
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-telegram-bot==13.14
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tornado==6.1
tzlocal==5.2
emoji==2.8.0